pyproject.toml
[project]
name = "ai-code-reviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "langchain-google-genai>=4.2.0",
    "openai>=2.14.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
]

[tool.hatchling.build.targets.wheel]
packages = ["src/reviewer"]
